Gathering

See merge request s12-fintech-finance-sub1/S12P21B110!135

+++ server/src/main/java/com/b110/jjeonchongmu/domain/mypage/service/MypageService.java
@@ -126,11 +126,13 @@ public class MypageService {
 
 		List<AutoPayment> autoPayments = autoPaymentRepo.findByUserId(user.getUserId());
 
-		List<GatheringProjection> gatheringProjections = customMyPageRepository.getAutoPaymentDtos(user.getUserId());
+		List<GatheringProjection> gatheringProjections = customMyPageRepository.getAutoPaymentDtos(
+				user.getUserId());
 		List<AutoPaymentDto> autoPaymentDtos = autoPayments.stream().flatMap(
 				autoPayment -> {
 					return gatheringProjections.stream()
-							.filter(projection -> projection.getGatheringId().equals(autoPayment.getGatheringAccount().getGathering().getGatheringId()))
+							.filter(projection -> projection.getGatheringId()
+									.equals(autoPayment.getGatheringAccount().getGathering().getGatheringId()))
 							.map(matchedProjection -> new AutoPaymentDto(
 									autoPayment.getAutoPaymentId(),
 									matchedProjection.getBasicFee(),
@@ -163,7 +165,7 @@ public class MypageService {
 			throw new RuntimeException("유저가 자동이체의 주인이 아닙니다.");
 		}
 
-		int amount = requestDto.getAmount();
+		long amount = requestDto.getAmount();
 		int day = requestDto.getDay();
 		boolean status = false;
 		if (requestDto.getStatus().equals("active")) {
